Cast uses of Layer and ViewportIndex to the expected type.

+++ reference/opt/shaders-msl/vert/viewport-index.msl2.invalid.vert
@@ -18,7 +18,7 @@ vertex main0_out main0(main0_in in [[stage_in]])
 {
     main0_out out = {};
     out.gl_Position = in.coord;
-    out.gl_ViewportIndex = int(in.coord.z);
+    out.gl_ViewportIndex = uint(int(in.coord.z));
     return out;
 }
